+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+ - Removed CallKit for china.
+
 ## [1.0.1 (7)] - 2020-11-17
 
  - We are now integrating with CallKit!